separate pageList
