+++ trelby/scenereport.py
@@ -4,6 +4,7 @@ import trelby.pml as pml
 import trelby.screenplay as screenplay
 import trelby.util as util
 
+from trelby.pagelist import PageList
 
 class SceneReport:
     def __init__(self, sp):
@@ -85,7 +86,7 @@ class SceneInfo:
         self.actionLines = 0
 
         # page numbers
-        self.pages = screenplay.PageList(sp.getPageNumbers())
+        self.pages = PageList(sp.getPageNumbers())
 
         # key = character name (upper cased), value = number of dialogue
         # lines
